Didnt do much. Giving up for the day
